migrating to authentication feature of supabase, add new folder and move dashboard setup and replace login page in main page.js, creating switching login/signup component and logout component, in supabase disabling confirm email in signup, adding fetching, inserting and other databse api calls, remove manual rls policy and add 1 each database table new auth rls policy for flexibility

+++ src/app/lib/permaDeleteDebtor.js
@@ -1,14 +1,16 @@
 'use client'
 
-import { supabase } from "./supabase";
+import { supabase } from "@/app/lib/supabase";
 
 export default function usePermaDeleteDebtor() {
 
     const deletePermanent = async (id) => {
 
+        const { data: { user } } = await supabase.auth.getUser();
         const {data, error} = await supabase
             .from('debtors')
             .delete()
+            .eq("user_id", user.id)
             .eq('id', id)
 
         if(error) throw error
